pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "extracttext"
version = "0.1.0"
description = "Unified text extraction for PDFs, images, DOCX, CSV, TXT"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [ { name = "Virgil Iordan", email = "iordan.virgil@gmail.com" } ]
keywords = ["ocr", "pdf", "text-extraction"]

dependencies = [
    "pdfminer.six>=20221105",
    "pdf2image>=1.16.2",
    "pytesseract>=0.3.10",
    "python-docx>=0.8.11",
    "Pillow>=10.0.0",
    "python-magic>=0.4.27",
    "chardet>=5.2.0"
]

[project.optional-dependencies]
ocr = ["pytesseract", "pdf2image", "Pillow", "poppler-utils"]

[project.scripts]
extracttext = "extracttext.cli:main"

[tool.setuptools]
packages = ["extracttext"] 